Set up basic data reading structure
